allow selecting post processes from advanced options

+++ crogine/src/graphics/SimpleDrawable.cpp
@@ -232,7 +232,7 @@ void SimpleDrawable::setVertexData(const std::vector<Vertex2D>& vertexData)
     glCheck(glBufferData(GL_ARRAY_BUFFER, vertexData.size() * sizeof(Vertex2D), vertexData.data(), GL_STATIC_DRAW));
     glCheck(glBindBuffer(GL_ARRAY_BUFFER, 0));
 
-    m_vertexCount = vertexData.size();
+    m_vertexCount = static_cast<std::uint32_t>(vertexData.size());
 }
 
 void SimpleDrawable::drawGeometry(const glm::mat4& worldTransform) const
